Fix rpath handling once and for all.
StackOfTasks.so (hrpsys plug-in) was missing symbols before this
commit. This is no longer the case.

This package now install libraries with their original rpath to make
sure library look-up is not broken by the CMake install mechanism.

The missing symbols were located in libplugin.so located in
${GRX_PREFIX}/${ROBOT}/lib. Plug-in is now linked explicitly against
this library.

Practically speaking, it means that starting from this commit, if
StackOfTasks.so is missing symbols, there plug-in will _fail_ and you
should recompile the control framework to fix the issue.

This will ease debug and make the compilation chain more robust.